uint64_t pour les Grid.content[i][j].state

+++ cell.h
@@ -1,11 +1,11 @@
-
+#include <stdint.h>
 #ifndef CELL_H
 #define CELL_H
 
 
 typedef struct {
     int x, y;
-    int state; // 1 si vivante, 0 si morte
+    uint64_t state; // 1 si vivante, 0 si morte
 }cell;
 
 int get_cell_state(cell Cell);
